feat(landing): :zap: add improvements
