+++ src/components/Cards/Rounded/index.tsx
@@ -32,8 +32,10 @@ const RoundedCard = (props: CardRounded) => {
           showLoadingAnimation={true}
           firstLaunchOnly={true}
           ready={false}
-          style={{ marginRight: '0', position: 'absolute' }}>
-        </ReactPlaceholder>
+          style={{ marginRight: '0', position: 'absolute' }}
+          // eslint-disable-next-line react/no-children-prop
+          children={null}
+        />
         <Image
           src={imageUrl}
           layout="fill"
